Change cli_options dir structure.
